Move ReactAriaRouterProvider out into shared webapp
