+++ application/account-management/WebApp/pages/__root.tsx
@@ -1,8 +1,8 @@
 import { createRootRoute, Outlet, useNavigate } from "@tanstack/react-router";
 import { ErrorPage } from "./-components/ErrorPage";
 import { NotFound } from "./-components/NotFoundPage";
-import { ReactAriaRouterProvider } from "@/shared/lib/router/ReactAriaRouterProvider";
 import { AuthenticationProvider } from "@/shared/lib/auth/AuthenticationProvider";
+import { ReactAriaRouterProvider } from "@repo/infrastructure/router/ReactAriaRouterProvider";
 
 export const Route = createRootRoute({
   component: Root,
