+++ application/account-management/WebApp/shared/lib/router/router.tsx
@@ -7,8 +7,6 @@ export const router = createRouter({
   defaultPreload: "intent"
 });
 
-export type RouterType = typeof router;
-
 // Register router with tanstack/react-router
 declare module "@tanstack/react-router" {
   interface Register {
